# openapi/person-api.yaml
openapi: 3.1.1
info:
  title: Person API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /people/{personId}:
    parameters:
      - name: personId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPerson
      summary: Get a person
      responses:
        '200':
          description: The stored person
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonUpdate'
        '404':
          description: No person exists for this identifier
    put:
      operationId: replacePerson
      summary: Replace a person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonUpdate'
      responses:
        '200':
          description: The saved person
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonUpdate'
        '400':
          description: The request body does not match the marital-status rules
components:
  schemas:
    Identity:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
      required:
        - firstName
        - lastName
    MarriedPerson:
      type: object
      properties:
        maritalStatus:
          const: married
        spouseFirstName:
          type: string
          minLength: 1
        spouseLastName:
          type: string
          minLength: 1
        spouseEmail:
          type: string
          minLength: 1
          pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$'
      required:
        - maritalStatus
        - spouseFirstName
        - spouseLastName
        - spouseEmail
    UnmarriedPerson:
      type: object
      properties:
        maritalStatus:
          enum:
            - single
            - divorced
            - widowed
      required:
        - maritalStatus
      not:
        anyOf:
          - required:
              - spouseFirstName
          - required:
              - spouseLastName
    PersonalInformation:
      allOf:
        - $ref: '#/components/schemas/Identity'
        - oneOf:
            - $ref: '#/components/schemas/MarriedPerson'
            - $ref: '#/components/schemas/UnmarriedPerson'
      discriminator:
        propertyName: maritalStatus
        mapping:
          married: '#/components/schemas/MarriedPerson'
          single: '#/components/schemas/UnmarriedPerson'
          divorced: '#/components/schemas/UnmarriedPerson'
          widowed: '#/components/schemas/UnmarriedPerson'
    Address:
      type: object
      properties:
        addressFirstLine:
          type: string
          minLength: 1
        addressSecondLine:
          type: string
        postCode:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
      required:
        - addressFirstLine
        - addressSecondLine
        - postCode
        - city
        - country
    EmergencyContact:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        relationship:
          type: string
        phoneNumber:
          type: string
          minLength: 1
        email:
          type: string
          pattern: '^$|^[^\s@]+@[^\s@]+\.[^\s@]+$'
      required:
        - name
        - relationship
        - phoneNumber
        - email
    EmergencyContactsDeclined:
      type: object
      properties:
        status:
          const: declined
      required:
        - status
    EmergencyContactsProvided:
      type: object
      properties:
        status:
          const: provided
        primaryEmergencyContact:
          $ref: '#/components/schemas/EmergencyContact'
        alternativeEmergencyContacts:
          type: array
          maxItems: 2
          items:
            $ref: '#/components/schemas/EmergencyContact'
      required:
        - status
        - primaryEmergencyContact
        - alternativeEmergencyContacts
    EmergencyContacts:
      oneOf:
        - $ref: '#/components/schemas/EmergencyContactsDeclined'
        - $ref: '#/components/schemas/EmergencyContactsProvided'
      discriminator:
        propertyName: status
        mapping:
          declined: '#/components/schemas/EmergencyContactsDeclined'
          provided: '#/components/schemas/EmergencyContactsProvided'
    PersonUpdate:
      type: object
      properties:
        personalInformation:
          $ref: '#/components/schemas/PersonalInformation'
        address:
          $ref: '#/components/schemas/Address'
        emergencyContacts:
          $ref: '#/components/schemas/EmergencyContacts'
      required:
        - personalInformation
        - address
        - emergencyContacts